requirements.txt
python-dotenv
fastapi
uvicorn
tiktoken
gunicorn
openai